requirements.txt
boto3
google-generativeai
psycopg2-binary
python-dotenv

